Fixed TimelineEvent Type file to not depend on the serverside DB client. This was causing problem with server deployment when used in lambda functions
